Refactoring useHubClusterName and adding documentation
